auth_service/requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.16
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
docling==2.45.0
docling-core==2.44.2
docling-ibm-models==3.9.0
docling-parse==4.1.0
durationpy==0.10
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gkeepapi==0.15.1
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gpsoauth==2.0.0
greenlet==3.2.4
griffe==1.11.1
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonlines==3.1.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==33.1.0
lark==1.2.2
latex2mathml==3.78.0
lazy_loader==0.4
llama-index-core==0.13.1
llama-index-embeddings-huggingface==0.6.0
llama-index-embeddings-openai==0.5.0
llama-index-instrumentation==0.4.0
llama-index-readers-google==0.7.0
llama-index-vector-stores-chroma==0.5.0
llama-index-workflows==1.3.0
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.99.9
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDrive==1.3.1
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semchunk==2.2.2
Send2Trash==1.8.3
sentence-transformers==5.1.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typer==0.16.0
types-python-dateutil==2.9.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0